Workspace & robot updates
This is actually Dave Z making these changes on the lab PC. Will was
logged in, so this will show up as his changes, but don't hold him
responsible :-)

Okay - big update.

Made quite a few changes to robotutil to update how the robot
initializes and moves.

Also moved to a different camera interface, using pyicic, a wrapper
around the Imaging Source drivers. Ultimately, we should actually pull
their code in programmatically, but for now, it's just copied in to our
codebase.

Added the concept of workspaces and modules, and created the first
module class files.

Finally, cleaned up the repository to remove test image files and other
junk. Added .bmp files to .gitignore so they won't get checked in in the
future.