Camera init update, plus driving the robot
Turns out the DLL loading has to match the version of Python (32- or
64-bit), regardless of the nature of the OS. So even a 64-bit operating
system requires a 32-bit DLL, if the Python version is 32-bit. Updated
the code to detect the Python version and match that.

Updated the SantaFeGUI.py program to effectively drive the robot around
the workspace. Controls X, Y, and Z0, Z1, and Z2.