Extents checking
Added configuration variables for the extents of the workspace, and also
for the offset between the different Z axes.

The move() family of functions now checks to make sure the coordinates
aren't beyond the extents (or less than zero), to prevent collisions or
hanging.

Note that it is still possible to crash the Z axes, since the modules
will be different heights and we can't account for that in the error
checking.