Updating config file setup
Changing the name of the configuration file to SantaFe.cfg. I'm deleting
the one in the repository. It's expected that at each different
computer, we'll make a copy of FlySorter.cfg (calling it SantaFe.cfg)
and edit that file. The repository is now set up to ignore SantaFe.cfg,
so you can easily make changes to that file that only impact your local
setup (min/max extents, heights, offsets, etc.).